librbd: deep_copy: move set_head to snapshot_copy state machine
