+++ src/librbd/deep_copy/SnapshotCopyRequest.cc
@@ -2,6 +2,7 @@
 // vim: ts=8 sw=2 smarttab
 
 #include "SnapshotCopyRequest.h"
+#include "SetHeadRequest.h"
 #include "SnapshotCreateRequest.h"
 #include "common/errno.h"
 #include "common/WorkQueue.h"
@@ -36,6 +37,7 @@ const std::string &get_snapshot_name(I *image_ctx, librados::snap_t snap_id) {
 
 } // anonymous namespace
 
+using librbd::util::create_context_callback;
 using librbd::util::unique_lock_name;
 
 template <typename I>
@@ -461,7 +463,7 @@ void SnapshotCopyRequest<I>::send_snap_protect() {
   if (snap_id_it == m_src_snap_ids.end()) {
     // no destination snapshots to protect
     m_prev_snap_id = CEPH_NOSNAP;
-    finish(0);
+    send_set_head();
     return;
   }
 
@@ -502,6 +504,45 @@ void SnapshotCopyRequest<I>::handle_snap_protect(int r) {
   }
 
   send_snap_protect();
+}
+
+template <typename I>
+void SnapshotCopyRequest<I>::send_set_head() {
+  if (m_snap_id_end != CEPH_NOSNAP) {
+    finish(0);
+    return;
+  }
+
+  ldout(m_cct, 20) << dendl;
+
+  uint64_t size;
+  ParentSpec parent_spec;
+  uint64_t parent_overlap;
+  {
+    RWLock::RLocker src_locker(m_src_image_ctx->snap_lock);
+    size = m_src_image_ctx->size;
+    parent_spec = m_src_image_ctx->parent_md.spec;
+    parent_overlap = m_src_image_ctx->parent_md.overlap;
+  }
+
+  auto ctx = create_context_callback<
+    SnapshotCopyRequest<I>, &SnapshotCopyRequest<I>::handle_set_head>(this);
+  auto req = SetHeadRequest<I>::create(m_dst_image_ctx, size, parent_spec,
+                                       parent_overlap, ctx);
+  req->send();
+}
+
+template <typename I>
+void SnapshotCopyRequest<I>::handle_set_head(int r) {
+  ldout(m_cct, 20) << "r=" << r << dendl;
+
+  if (r < 0) {
+    lderr(m_cct) << "failed to set head: " << cpp_strerror(r) << dendl;
+    finish(r);
+    return;
+  }
+
+  finish(0);
 }
 
 template <typename I>
